feat: 딜러(Dealer) 구현
