+++ README.md
@@ -57,10 +57,10 @@
 ### 기능 구현 목록
 
 - 딜러 (Dealer)
-    - [ ] 기본 구성은 플레이어와 같다.
-    - [ ] 카드를 추가로 뽑을 수 있는 상황인지 확인할 수 있다.
-        - [ ] 처음에 받은 2장의 합계가 16이하이면 반드시 1장의 카드를 추가한다.
-        - [ ] 카드의 총 점수가 17점 이상이면 추가로 받을 수 없다.
+    - [x] 기본 구성은 플레이어와 같다.
+    - [x] 카드를 추가로 뽑을 수 있는 상황인지 확인할 수 있다.
+        - [x] 처음에 받은 2장의 합계가 16이하이면 반드시 1장의 카드를 추가한다.
+        - [x] 카드의 총 점수가 17점 이상이면 추가로 받을 수 없다.
 
 
 - 게임 결과 (GameResult)
