+++ src/main/kotlin/blackjack/domain/Player.kt
@@ -1,24 +1,11 @@
 package blackjack.domain
 
-class Player(private val name: PlayerName, private val hand: Hand) {
-    fun addCard(card: Card) {
-        hand.addCard(card)
-    }
-
-    fun displayHand(): String {
-        return hand.getCards().joinToString(", ") { it.display() }
-    }
-
-    fun isDrawable(): Boolean {
+class Player(
+    private val name: PlayerName,
+    private val hand: Hand,
+) : Participant(name, hand) {
+    override fun isDrawable(): Boolean {
         return hand.calculateBestTotal() != 0
-    }
-
-    fun calculateTotal(): Int {
-        return hand.calculateBestTotal()
-    }
-
-    fun getName(): String {
-        return name.value
     }
 
     companion object {
